+++ bmcmanager/commands/ipmi/ssh.py
@@ -18,11 +18,13 @@ from bmcmanager.commands.base import BMCManagerServerCommand
 
 class SSH(BMCManagerServerCommand):
     """
-    Connect to IPMI using SSH
+    connect to IPMI using SSH
     """
     oob_method = 'ipmi_ssh'
 
     def get_parser(self, prog_name):
         parser = super().get_parser(prog_name)
-        parser.add_argument('command', nargs='*')
+        parser.add_argument(
+            'command', nargs='*',
+            help='execute this command instead of starting a shell')
         return parser