+++ bmcmanager/commands/open.py
@@ -18,7 +18,7 @@ from bmcmanager.commands.base import BMCManagerServerCommand
 
 class DCIM(BMCManagerServerCommand):
     """
-    Open DCIM
+    open server page on DCIM
     """
     dcim_fetch_secrets = False
     oob_method = 'open_dcim'
@@ -26,7 +26,7 @@ class DCIM(BMCManagerServerCommand):
 
 class Web(BMCManagerServerCommand):
     """
-    Open IPMI Web interface
+    open IPMI web interface
     """
     oob_method = 'open'
     dcim_fetch_secrets = False
@@ -34,11 +34,13 @@ class Web(BMCManagerServerCommand):
 
 class Console(BMCManagerServerCommand):
     """
-    Open Console
+    open console
     """
     oob_method = 'console'
 
     def get_parser(self, prog_name):
         parser = super().get_parser(prog_name)
-        parser.add_argument('--print-cmd', action='store_true', default=False)
+        parser.add_argument(
+            '--print-cmd', action='store_true', default=False,
+            help='orint the command instead of opening the console')
         return parser