+++ bmcmanager/commands/ram.py
@@ -18,18 +18,20 @@ from bmcmanager.commands.base import BMCManagerServerCommand, BMCManagerServerGe
 
 class Get(BMCManagerServerGetCommand):
     """
-    Print server RAM
+    print server RAM
     """
     oob_method = 'system_ram'
 
 
 class Check(BMCManagerServerCommand):
     """
-    Nagios check for server RAM
+    check server RAM [Nagios]
     """
     oob_method = 'check_ram'
 
     def get_parser(self, prog_name):
         parser = super().get_parser(prog_name)
-        parser.add_argument('--expected', type=int, default=None)
+        parser.add_argument(
+            '--expected', type=int, default=None,
+            help='gb of RAM the server should have')
         return parser